Move role-based actions to submodule
This change moves the role-based actions to the submodule,
actions::rbac::roles. The new_client function remains in the base
module, as it will be used in other submodules.
